+++ core/main.py
@@ -525,13 +525,9 @@ class MainWindow(QMainWindow, Ui_MainWindow):
     def setLinkageFreemove(self, enable: bool):
         _entities.setLinkageFreemove(self, enable)
     
-    @pyqtSlot(float)
-    def on_linkage_freemode_slider_valueChanged(self, value: float):
+    @pyqtSlot(int)
+    def on_linkage_freemode_slider_valueChanged(self, value: int):
         _entities.on_linkage_freemode_slider_valueChanged(self, value)
-    
-    @pyqtSlot()
-    def on_linkage_freemode_slider_sliderReleased(self):
-        _entities.on_linkage_freemode_slider_sliderReleased(self)
     
     def setCoordsAsCurrent(self):
         _entities.setCoordsAsCurrent(self)