+++ changelog.md
@@ -8,6 +8,7 @@ Features
 ---
 
 + Free move mode:
+    + Linkage editing is supported in expression table.
     + Fix the error of angle updating.
 + Show the values on expression table.
 + Add solution selection mode with expression table.